+++ spec/document/elements/paragraph/runs/drawing/properties/drawing_properties_color_spec.rb
@@ -7,7 +7,7 @@ describe 'Drawing Properties Color' do
     expect(docx.elements.first.character_style_array[0]
                .alternate_content.office2010_content.graphic.data.properties.fill_color.type).to eq(:gradient)
     expect(docx.elements.first.character_style_array[0]
-               .alternate_content.office2010_content.graphic.data.properties.fill_color.value.gradient_stops[0].color).to eq(OoxmlParser::Color.new(235, 241, 221))
+               .alternate_content.office2010_content.graphic.data.properties.fill_color.value.gradient_stops[0].color.value).to eq(OoxmlParser::Color.new(235, 241, 221))
     expect(docx.elements.first.character_style_array[0]
                .alternate_content.office2010_content.graphic.data.properties.fill_color.value
                .gradient_stops[0].position - 10).to be < 3
@@ -35,5 +35,17 @@ describe 'Drawing Properties Color' do
     expect(docx.element_by_description(location: :canvas, type: :paragraph).first.nonempty_runs.first
                .alternate_content.office2010_content.graphic.data.properties
                .fill_color.value.background_color).to eq(OoxmlParser::Color.new(237, 125, 49))
+  end
+
+  it 'shape_gradient_color_angle.docx' do
+    docx = OoxmlParser::DocxParser.parse_docx('spec/document/elements/paragraph/runs/drawing/properties/colors/shape_gradient_color_angle.docx')
+    expect(docx.elements.first.character_style_array[1]
+               .alternate_content.office2010_content.graphic.data.properties.fill_color.value.linear_gradient.angle).to eq(13.37)
+  end
+
+  it 'graphic_background_color_gradient.docx' do
+    docx = OoxmlParser::DocxParser.parse_docx('spec/document/elements/paragraph/runs/drawing/properties/colors/graphic_background_color_gradient.docx')
+    drawing = docx.element_by_description[1].nonempty_runs.first.drawing
+    expect(drawing.graphic.data.shape_properties.fill.value.gradient_stops[0].color.converted_color).to eq(OoxmlParser::Color.new(237, 237, 237))
   end
 end